Make some tweaks to how we detect generated files
I've seen occasional flakiness in our testing of this feature, but
couldn't pinpoint the reason for it. However, decreasing the frequency
how often we check for changes of the `command.log` and looking at a
larger chunk of the file seems to help (based on running th tests
multiple times).


Bug: http://b/184120864
Change-Id: Ie28a6e5aa71e7d86e0afc47906051ee1c05ed44c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/194786
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Michal Terepeta <michalt@google.com>