[analyzer] Improve the detection of file changes in bazel_watcher
Previously we used timestamps to detect if a file has been changed.
This works for files that rarely change. However, for a file that is
changed a lot, this doesn't really work reliably due to the low
resolution of timestamps. This is the case for `command.log` that we
use to detect Bazel builds: Bazel writes to it continuously so we
detect a lot of changes, but the last update to the file might end up
having the same timestamp, which means that we might miss the fact
that a build finished.

This commit changes the code to also use the length of the file when
trying to figure out if a file has been changed. A better check would
be to hash the contents, but I'm slightly afraid of the performance
impact for files like `command.log` that we watch, and which can be
both fairly large (multiple megabytes) and change a lot.  Even for
other files, if we have a large build that finished, we could have
1000s of file that we'd need to read and hash. So for now I went with
the size of the file instead. I could be convinced otherwise though ;)

Bug: http://b/184120864 and http://b/186200325

Change-Id: I1bf13adab5164c9ec04b59b275f2c83b96d08ea4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/197043
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Michal Terepeta <michalt@google.com>