[ddc] remove extra call of `addTypeCaches`
The caches on FutureOr types were accidentally being cleared every time
one was created with the same type argument. These caches are already
added on every FutureOr class that is created in the wrapping for
generic types.

This bug surfaced when a class with a circular type hierarchy that must
be deferred included a FutureOr. The result was infinite recursion when
creating the type because the caches were always empty.

This change showed no notable differences in performance:
Weak:
https://golem.corp.goog/Comparison?repository=dart&team=dartdevc#targetA%3Ddartdevc%3BmachineTypeA%3Dlinux-x64%3BrevisionA%3D92060%3BpatchA%3Dnshahan-FutureOr-perf%3BtargetB%3Ddartdevc%3BmachineTypeB%3Dlinux-x64%3BrevisionB%3D92059%3BpatchB%3DNone
Sound:
https://golem.corp.goog/Comparison?repository=dart&team=dartdevc#targetA%3Ddartdevc-null%3BmachineTypeA%3Dlinux-x64%3BrevisionA%3D92060%3BpatchA%3Dnshahan-FutureOr-perf%3BtargetB%3Ddartdevc-null%3BmachineTypeB%3Dlinux-x64%3BrevisionB%3D92058%3BpatchB%3DNone

Change-Id: I985e0ab2898a8a68ac6cdf4d0a738120bc42c623
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/197900
Reviewed-by: Mark Zhou <markzipan@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>