[ddc] Add the prefix '__t$' to the names of types
This is a temporary fix to give debug tools a recognizable string so
they can hide the type variables from the current scope.

The prefix is only applied when the type appears outside of the
top level type table.

This will require a change in dwds to filter variables with names that
begin with this string.

Change-Id: I2942f462280195d81a7de2b710e6b958770d152a
Issue: https://github.com/dart-lang/sdk/issues/45649
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/199581
Reviewed-by: Anna Gringauze <annagrin@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>